Move favorites controller and stub checkout
