+++ src/main/webapp/WEB-INF/web.xml
@@ -153,7 +153,7 @@
     </servlet>
     <servlet-mapping>
         <servlet-name>FavoriteController</servlet-name>
-        <url-pattern>/favoritos/*</url-pattern>
+        <url-pattern>/favorites/*</url-pattern>
     </servlet-mapping>
     
     <!-- Administração -->
